# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blueberry AI Elastic Tools",
    "description": "API tools for querying and updating customer retention data stored in Elasticsearch Serverless.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://blueberry-ai-359524452928.us-central1.run.app",
      "description": "Production Cloud Run server"
    }
  ],
  "paths": {
    "/api/tools/account-context": {
      "get": {
        "summary": "Get Account Context",
        "description": "Fetches complete customer details, open tickets, recent CSM health notes, and call transcripts from Elasticsearch.",
        "operationId": "getAccountContext",
        "parameters": [
          {
            "name": "accountId",
            "in": "query",
            "description": "The unique identifier of the customer account (e.g. ACC-002)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved account context",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "account": {
                      "type": "object",
                      "properties": {
                        "account_id": { "type": "string" },
                        "company_name": { "type": "string" },
                        "industry": { "type": "string" },
                        "arr": { "type": "number" },
                        "risk_score": { "type": "number" },
                        "status": { "type": "string" },
                        "last_contact_date": { "type": "string" }
                      }
                    },
                    "tickets": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "ticket_id": { "type": "string" },
                          "account_id": { "type": "string" },
                          "subject": { "type": "string" },
                          "description": { "type": "string" },
                          "status": { "type": "string" },
                          "priority": { "type": "string" },
                          "created_at": { "type": "string" }
                        }
                      }
                    },
                    "healthNotes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "note_id": { "type": "string" },
                          "account_id": { "type": "string" },
                          "author": { "type": "string" },
                          "note_text": { "type": "string" },
                          "sentiment": { "type": "string" },
                          "created_at": { "type": "string" }
                        }
                      }
                    },
                    "callTranscripts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "call_id": { "type": "string" },
                          "account_id": { "type": "string" },
                          "transcript": { "type": "string" },
                          "summary": { "type": "string" },
                          "duration_minutes": { "type": "integer" },
                          "date": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "404": {
            "description": "Account not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/tools/search-issues": {
      "get": {
        "summary": "Search Customer Issues",
        "description": "Searches for matching tickets, notes, and transcripts semantically using a natural language query.",
        "operationId": "searchIssues",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "description": "The natural language search query (e.g. 'SSO login error' or 'threatened to churn')",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "in": "query",
            "description": "Optional account ID to filter search results to a specific customer",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tickets": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_score": { "type": "number" },
                          "ticket_id": { "type": "string" },
                          "account_id": { "type": "string" },
                          "subject": { "type": "string" },
                          "description": { "type": "string" },
                          "status": { "type": "string" },
                          "priority": { "type": "string" }
                        }
                      }
                    },
                    "transcripts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_score": { "type": "number" },
                          "call_id": { "type": "string" },
                          "account_id": { "type": "string" },
                          "transcript": { "type": "string" },
                          "summary": { "type": "string" }
                        }
                      }
                    },
                    "healthNotes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_score": { "type": "number" },
                          "note_id": { "type": "string" },
                          "account_id": { "type": "string" },
                          "note_text": { "type": "string" },
                          "sentiment": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/tools/write-note": {
      "post": {
        "summary": "Write Health Note",
        "description": "Adds a new customer health note or escalation summary into Elasticsearch.",
        "operationId": "writeHealthNote",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "accountId",
                  "noteText"
                ],
                "properties": {
                  "accountId": {
                    "type": "string",
                    "description": "The unique identifier of the customer account (e.g. ACC-002)"
                  },
                  "noteText": {
                    "type": "string",
                    "description": "The content of the note or escalation summary"
                  },
                  "sentiment": {
                    "type": "string",
                    "description": "The sentiment associated with the note",
                    "enum": [
                      "Positive",
                      "Neutral",
                      "Negative"
                    ]
                  },
                  "author": {
                    "type": "string",
                    "description": "The author of the note",
                    "default": "Blueberry Copilot"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully indexed health note",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "note": { "type": "object" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/tools/agent-memory": {
      "get": {
        "summary": "Get Agent Memory",
        "description": "Retrieves the long-term agent memories and notes remembered about a specific account.",
        "operationId": "getAgentMemory",
        "parameters": [
          {
            "name": "accountId",
            "in": "query",
            "description": "The unique identifier of the customer account (e.g. ACC-002)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "in": "query",
            "description": "Optional semantic query to retrieve specific memories matching a topic",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved agent memories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "memories": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_score": { "type": "number" },
                          "memory_id": { "type": "string" },
                          "account_id": { "type": "string" },
                          "content": { "type": "string" },
                          "category": { "type": "string" },
                          "created_at": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Write Agent Memory",
        "description": "Saves a key fact, preference, or escalation milestone to the long-term agent memory index.",
        "operationId": "writeAgentMemory",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "accountId",
                  "content"
                ],
                "properties": {
                  "accountId": {
                    "type": "string",
                    "description": "The unique identifier of the customer account"
                  },
                  "content": {
                    "type": "string",
                    "description": "The fact or preference to remember"
                  },
                  "category": {
                    "type": "string",
                    "description": "Category of memory (preference, escalation, milestone)",
                    "default": "milestone"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully saved agent memory",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "memory": { "type": "object" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/tools/dynamic-risk": {
      "get": {
        "summary": "Calculate Dynamic Churn Risk",
        "description": "Uses live ES|QL aggregations to dynamically calculate the retention risk of a customer account and write it back to Elasticsearch.",
        "operationId": "getDynamicRiskScore",
        "parameters": [
          {
            "name": "accountId",
            "in": "query",
            "description": "The unique identifier of the customer account (e.g. ACC-002)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dynamic risk calculation complete"
          }
        }
      }
    },
    "/api/tools/recommend-runbook": {
      "get": {
        "summary": "Recommend Resolution Runbook",
        "description": "Semantically searches the knowledge base index to match runbook steps for a ticket description or search string.",
        "operationId": "recommendRunbook",
        "parameters": [
          {
            "name": "ticketId",
            "in": "query",
            "description": "Optional support ticket ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "in": "query",
            "description": "Optional raw search string",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved runbook recommendations"
          }
        }
      }
    },
    "/api/tools/escalate": {
      "post": {
        "summary": "Trigger Account Escalation",
        "description": "Sets account risk to 99%, creates an escalation milestone in agent memory, and generates Slack JSON block kit cards and email templates.",
        "operationId": "escalateAccount",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["accountId"],
                "properties": {
                  "accountId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Escalation triggered successfully"
          }
        }
      }
    },
    "/api/tools/simulate-event": {
      "post": {
        "summary": "Simulate Customer Event Ingestion",
        "description": "Indexes a new ticket, call, or CSM note to simulate customer activity, automatically updating risk metrics.",
        "operationId": "simulateEvent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["type", "accountId"],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": ["ticket", "call", "note"]
                  },
                  "accountId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event indexed and metrics re-computed"
          }
        }
      }
    },
...
  }
}